refactor: move the mediator into a top-level Messaging folder
