+++ OLED-Sleeper.Tests/Features/MonitorBehavior/Handlers/ApplyMonitorIdleBehaviorCommandHandlerTests.cs
@@ -1,5 +1,4 @@
 using Moq;
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Features.MonitorBehavior.Commands;
 using OLED_Sleeper.Features.MonitorBehavior.Handlers;
 using OLED_Sleeper.Features.MonitorBehavior.Models;
@@ -7,6 +6,7 @@ using OLED_Sleeper.Features.MonitorBlackout.Commands;
 using OLED_Sleeper.Features.MonitorDimming.Commands;
 using OLED_Sleeper.Features.MonitorIdleDetection.Models;
 using OLED_Sleeper.Features.UserSettings.Models;
+using OLED_Sleeper.Messaging.Interfaces;
 using System.Windows;
 
 namespace OLED_Sleeper.Tests.Features.MonitorBehavior.Handlers
