+++ OLED-Sleeper/Features/MonitorBehavior/Commands/ApplyMonitorActiveBehaviorCommand.cs
@@ -1,5 +1,5 @@
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Features.MonitorIdleDetection.Models;
+using OLED_Sleeper.Messaging.Interfaces;
 
 namespace OLED_Sleeper.Features.MonitorBehavior.Commands
 {
