+++ OLED-Sleeper/Features/MonitorDimming/Handlers/RestoreBrightnessOnAllMonitorsCommandHandler.cs
@@ -1,6 +1,6 @@
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Features.MonitorDimming.Commands;
 using OLED_Sleeper.Features.MonitorDimming.Services.Interfaces;
+using OLED_Sleeper.Messaging.Interfaces;
 using Serilog;
 
 namespace OLED_Sleeper.Features.MonitorDimming.Handlers
