+++ OLED-Sleeper/Features/MonitorState/Commands/SynchronizeMonitorStateCommand.cs
@@ -1,5 +1,5 @@
-﻿using OLED_Sleeper.Core.Interfaces;
-using OLED_Sleeper.Features.MonitorInformation.Models;
+﻿using OLED_Sleeper.Features.MonitorInformation.Models;
+using OLED_Sleeper.Messaging.Interfaces;
 
 namespace OLED_Sleeper.Features.MonitorState.Commands
 {
