+++ OLED-Sleeper/Features/UserSettings/Commands/ApplySettingsChangeCommand.cs
@@ -1,5 +1,5 @@
-using OLED_Sleeper.Core.Interfaces;
 using OLED_Sleeper.Features.UserSettings.Models;
+using OLED_Sleeper.Messaging.Interfaces;
 
 namespace OLED_Sleeper.Features.UserSettings.Commands
 {
